backend/requirements/common.txt
django~=3.0.0
django-rest-framework
django-cors-headers
djangorestframework-jwt
pillow